ResNet18_K-fold_Data_Augmentation/inference.py
import torch
from torchvision import transforms
from PIL import Image

def inference_model(model, image_path, config):
    model.eval()
    transform = transforms.Compose([
        transforms.Resize(256),
        transforms.CenterCrop(224),
        transforms.ToTensor(),
        transforms.Normalize(mean=config['data']['mean'], std=config['data']['std']),
    ])

    image = Image.open(image_path).convert('RGB')
    image = transform(image).unsqueeze(0)

    with torch.no_grad():
        outputs = model(image)
        _, predicted = torch.max(outputs, 1)
        return predicted.cpu().numpy()

