ResNet18_K-fold_Data_Augmentation/requirements.txt
torch
torchvision
tqdm
opencv-python
scikit-learn
matplotlib
seaborn
pillow
torchsampler
pyyaml
